Add home page on-call card
