+++ incident/src/components/HomePageOnCallCard/index.ts
@@ -1,0 +1,3 @@
+export { Content } from "./Content";
+
+export const ContextProvider = ({ children }: { children: JSX.Element }) => children;
